+++ src/ua/data_types/data_value.rs
@@ -58,11 +58,10 @@ impl DataValue {
         self
     }
 
+    #[deprecated = "use Self::with_status() instead"]
     #[must_use]
-    pub fn with_status_code(mut self, status_code: &ua::StatusCode) -> Self {
-        status_code.clone_into_raw(&mut self.0.status);
-        self.0.set_hasStatus(true);
-        self
+    pub fn with_status_code(self, status_code: &ua::StatusCode) -> Self {
+        self.with_status(status_code)
     }
 
     /// Gets value.
@@ -105,10 +104,16 @@ impl DataValue {
     }
 
     #[must_use]
-    pub fn status_code(&self) -> Option<ua::StatusCode> {
+    pub fn status(&self) -> Option<ua::StatusCode> {
         self.0
             .hasStatus()
             .then(|| ua::StatusCode::new(self.0.status))
+    }
+
+    #[deprecated = "use Self::status() instead"]
+    #[must_use]
+    pub fn status_code(&self) -> Option<ua::StatusCode> {
+        self.status()
     }
 
     pub(crate) fn to_generic<T: DataType>(&self) -> Result<crate::DataValue<T>> {